test(library): add tests for reading contexts in type-safe way
